backend/requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bioc==2.1
boto3==1.38.19
botocore==1.38.19
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
conllu==4.5.3
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
flair==0.15.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
gdown==5.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
HyperPyYAML==1.2.2
idna==3.10
iniconfig==2.1.0
intervaltree==3.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
jsonlines==4.0.0
julius==0.2.7
kiwisolver==1.4.8
langdetect==1.0.9
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
openai==1.82.0
openai-whisper==20240930
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
pptree==3.1
primePy==1.3
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytorch_revgrad==0.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
segtok==1.5.11
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
speechbrain==1.0.3
SQLAlchemy==2.0.41
sqlitedict==2.1.0
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.2.2
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.1
tqdm==4.67.1
transformer-smaller-training-vocab==0.4.1
transformers==4.51.3
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
Wikipedia-API==0.8.1
wrapt==1.17.2
yarl==1.20.0
